==> Actualización/1. Procesamiento de proveedores.ipynb <==
import pandas as pd
import numpy as np
import re as re

La idea es tomar el dataframe de Proveedores, con todos los precios de los proveedores, y comparar para sacar el más barato. Con esto, dejamos un .csv nuevo que va a estar listo para usarse en el aumento de precios. 
La idea es que compare estos precios con los que se bajan en el exporte.

Siempre asegurémonos de eliminar símbolos y demás cuando pasamos precios al Excel de Proveedores, así queda todo bien, y no hay que convertir strings, etc.

Creamos el dataframe.

Proveedores = 'J:/My Drive/Forraje/Proveedores.xlsx'

# Crear el DataFrame utilizando esas columnas
df_Proveedores = pd.read_excel(Proveedores, sheet_name='Centro')


Ponemos las listas de columnas.

lista_proveedores = ['Benavídez', 'Gallo', 'Anea', 'Pilar', 'Tortuguitas', 'Del Viso',
                     'Chica Glade', 'San Martín', 'Hipocampus', 'Mundo Mascotas', 'Yaguar',
                     'King Clor', 'Diserquim', 'Moreno', 'Roberto', 'Viejo Gets', 'Viejo Cuero', 'Mis Nietos', 
                     'Profesional Vet', 'Premier', 'Dacam', 'Proveedor', 'Saphirus']

Convertimos a float los valores strings que puedan llegar a ser problemáticos. También borramos en el caso de que haya palabras.

# Convertir las columnas especificadas a float64
# for i in lista_proveedores:
#     df_Proveedores[i] = df_Proveedores[i].astype(float)


Calcula el mínimo entre todas las columnas, e ignora los cero. Si todas son cero, queda cero.

# Calcula el mínimo de las columnas especificadas, excluyendo los valores de cero
min_sin_cero = df_Proveedores[lista_proveedores].replace(0, np.nan).min(axis=1, skipna=True)


# Actualiza la columna 'Precios' con los mínimos calculados
df_Proveedores['Precio'] = min_sin_cero


Ahora lo que hace es llenar la columna "Proveedor" con el proveedor que tiene más barato el producto.

# Iteramos sobre cada fila del DataFrame
for indice, fila in df_Proveedores.iterrows():
    # Iteramos sobre cada columna de lista_proveedores
    for columna in lista_proveedores:
        # Comparamos el valor de "Precio" con el valor en la columna actual
        if fila["Precio"] == fila[columna]:
            # Si coinciden, actualizamos la columna "Proveedores" con el nombre de la columna actual
            df_Proveedores.at[indice, "Proveedor"] = columna
            # Salimos del bucle interno
            break


Creamos un nuevo dataframe con las columnas correspondientes.

# Crear el DataFrame utilizando esas columnas
df = pd.DataFrame()

Asignamos las columnas correspondientes al nuevo df.

df['Codigo'] = df_Proveedores['Codigo']
df['Descripcion'] = df_Proveedores['Descripcion']
df['Costo'] = df_Proveedores['Precio']
df['Categoria'] = df_Proveedores['Categoria']
df['Proveedor'] = df_Proveedores['Proveedor']

df.head()

Creamos un .csv con la fecha actual.

# Guardar el DataFrame en un archivo CSV con la fecha actual en el nombre
nombre_archivo = f"J:/My Drive/Forraje/Calculados/Costos comparados.csv"
df.to_csv(nombre_archivo, index=False)
